Reworked checkForWin() method, fixed issue #3
Reworked checkForWin() method to count the number of elements with show and hide classes and compare them to the length of the string instead of looking at each li and determining if it has a class of "show".  This gets more accurate results when determining if the player has guessed all of the letters in the phrase.